infrastructure/email/requirements.txt
# Core dependencies
boto3==1.28.84
redis==5.0.1
psycopg2-binary==2.9.9
flask==3.0.0
pydantic==2.5.0

# Async support
asyncio==3.4.3
aioredis==2.0.1

# Email processing
email-validator==2.1.0
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
pandas==2.1.4
schedule==1.2.0

# Security
cryptography==41.0.7

# HTTP client
requests==2.31.0
aiohttp==3.9.1